refactor dashboard: moved dashboard logic
