+++ src/app/dashboard/dashboard.html
@@ -1,4 +1,4 @@
-<mat-grid-list [cols]="cols()" rowHeight="160px" gutterSize="16px" class="board">
+<!-- <mat-grid-list [cols]="cols()" rowHeight="160px" gutterSize="16px" class="board">
   <mat-grid-tile *ngFor="let t of tiles()"
                  [colspan]="t.colspan"
                  [rowspan]="t.rowspan"
@@ -13,4 +13,4 @@
       </div>
     </div>
   </mat-grid-tile>
-</mat-grid-list>
+</mat-grid-list> -->
