+++ src/app/dashboard/widgets/registry.ts
@@ -1,21 +1,21 @@
-import { PluginDef } from './types';
+// import { PluginDef } from './types';
 
 
-//dummy data
-export const PLUGINS: PluginDef[] = [
-  {
-    id: 'mews',
-    name: 'Mews Widgets',
-    widgets: [
-      { id: 'reserveringen', title: 'reserveringen', description: 'Toon reserveringen' }
-    ],
-  },
-  {
-    id: 'robots',
-    name: 'robots',
-    widgets: [
-      { id: 'robotlijst', title: 'Robotlijst' },
-      { id: 'robotflows', title: 'robotflows' },
-    ],
-  },
-];
+// //dummy data
+// export const PLUGINS: PluginDef[] = [
+//   {
+//     id: 'mews',
+//     name: 'Mews Widgets',
+//     widgets: [
+//       { id: 'reserveringen', title: 'reserveringen', description: 'Toon reserveringen' }
+//     ],
+//   },
+//   {
+//     id: 'robots',
+//     name: 'robots',
+//     widgets: [
+//       { id: 'robotlijst', title: 'Robotlijst' },
+//       { id: 'robotflows', title: 'robotflows' },
+//     ],
+//   },
+// ];
