+++ src/app/dashboard/widgets/types.ts
@@ -1,11 +1,11 @@
-export interface WidgetDef {
-  id: string;        
-  title: string;    
-  description?: string;
-}
+// export interface WidgetDef {
+//   id: string;        
+//   title: string;    
+//   description?: string;
+// }
 
-export interface PluginDef {
-  id: string;        
-  name: string;     
-  widgets: WidgetDef[];
-}
+// export interface PluginDef {
+//   id: string;        
+//   name: string;     
+//   widgets: WidgetDef[];
+// }
